requirements.txt
charset-normalizer==3.0.1
Django==3.0.8
django-environ==0.9.0
djangorestframework==3.14.0
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
pygments